Fix TileEntity issue
Inventory expects onInventoryChanged() to be defined in classes that
implement it. TileEntity defines that method, but doesn't explicitly
implement Inventory. ChestTileEntity extends TileEntity and implements
Inventory, and the compiler is happy because it wants y_() and gets
y_() in an obfuscated environment, but we need to map both if we want
to deobfuscate it.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>